cpufreq-ondemand-Set-MIN_LATENCY_MULTIPLIER-to-20
Currently MIN_LATENCY_MULTIPLIER is set defined as 100 and so on a
system with
transition latency of 1 ms, the minimum sampling time comes to be around
100 ms.
That is quite big if you want to get better performance for your system.

Redefine MIN_LATENCY_MULTIPLIER to 20 so that we can support 20ms
sampling rate
for such platforms.

Signed-off-by: faux123 <reioux@gmail.com>
Signed-off-by: Simarpreet Singh <simar@linux.com>